Fix for Xcode 6.3 Beta 3

+++ XCGLogger/Library/XCGLoggerTests/XCGLoggerTests.swift
@@ -131,8 +131,8 @@ class XCGLoggerTests: XCTestCase {
 
         let linesToLog = ["One", "Two", "Three", "Four", "Five", "Six", "Seven", "Eight", "Nine", "Ten"]
         let myConcurrentQueue = dispatch_queue_create("com.cerebralgardens.xcglogger.testMultiThreaded.queue", DISPATCH_QUEUE_CONCURRENT)
-        dispatch_apply(UInt(linesToLog.count), myConcurrentQueue) { (index: UInt) in
-            log.debug(linesToLog[Int(index)])
+        dispatch_apply(linesToLog.count, myConcurrentQueue) { (index: Int) in
+            log.debug(linesToLog[index])
         }
     }
 }
